CXC-307 : Changes to mobile responsiveness style
Chnages tot search length, overlay layout modal, popupOverlay name and logic
